Adding redis cache
