+++ docker-compose.yaml
@@ -1,0 +1,19 @@
+version: '3.8'
+services:
+  #app:
+    #container_name: app
+    #image: app_image
+    #restart: always
+    #build:
+     #   context: .
+      ##  dockerfile: Dockerfile
+    ##   - "80:80"
+
+  app_redis:
+    image: redis:latest
+    hostname: redis
+    container_name: app_redis
+    command:
+      - 'redis-server'
+    ports:
+      - "6379:6379"
