+++ main.py
@@ -1,27 +1,21 @@
-from enum import Enum
 from typing import List
+
+from aiocache import cached, Cache
 from fastapi import FastAPI
-from pydantic import BaseModel, PositiveInt, PositiveFloat
+from fastapi.middleware.cors import CORSMiddleware
 
-
-class Status(Enum):
-    completed = "completed"
-    pending = "pending"
-    canceled = "canceled"
-    all = "all"
-
-
-class Order(BaseModel):
-    id: PositiveInt
-    item: str
-    quantity: PositiveInt
-    price: PositiveFloat
-    status: Status = Status.all
-
-
-LIST_ORDER_STATUS = dict[str, list[Order] | Status]
+from model.order import Order, Status, LIST_ORDER_STATUS
 
 app = FastAPI()
+
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=["*"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
 
 
 @app.get("/", tags=["App"])
@@ -34,10 +28,12 @@ async def say_hello(name: str):
     return {"message": f"Hello {name}"}
 
 
-@app.post("solution", tags=["App"])
-async def process_orders(orders: List[Order], criterion: Status)-> LIST_ORDER_STATUS:
+@app.post("solution", tags=["Solution"])
+@cached(ttl=3600, cache=Cache.REDIS, endpoint="localhost", port="6379")
+async def process_orders(orders: List[Order], criterion: Status) -> LIST_ORDER_STATUS:
     filtered_orders = []
     for order in orders:
         if (order.status == criterion) or (criterion == Status.all):
             filtered_orders.append(order)
     return {"orders": filtered_orders, "criterion": criterion}
+
